[refactor] : 예약 결제 비즈니스 로직 도메인으로 변경
