+++ src/main/java/kr/hhplus/be/server/payment/domain/service/PaymentService.java
@@ -1,6 +1,5 @@
 package kr.hhplus.be.server.payment.domain.service;
 
-import kr.hhplus.be.server.concert.domain.repository.SeatRepository;
 import kr.hhplus.be.server.payment.domain.model.Payment;
 import kr.hhplus.be.server.payment.domain.repository.PaymentRepository;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +11,6 @@ import org.springframework.stereotype.Service;
 public class PaymentService {
 
     private final PaymentRepository paymentRepository;
-    private final SeatRepository seatRepository;
 
     public Payment processPayment(Long reservationId, Long userId, Long amount) {
