+++ src/main/java/kr/hhplus/be/server/reservation/application/dto/PaymentResult.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.payment.application.dto;
+package kr.hhplus.be.server.reservation.application.dto;
 
 import kr.hhplus.be.server.payment.domain.model.Payment;
