+++ src/main/java/kr/hhplus/be/server/reservation/presentation/dto/response/PaymentResponse.java
@@ -1,10 +1,7 @@
-package kr.hhplus.be.server.payment.presentation.dto.response;
+package kr.hhplus.be.server.reservation.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.hhplus.be.server.concert.presentation.dto.response.SeatResponse;
-import kr.hhplus.be.server.payment.application.dto.PaymentResult;
-import kr.hhplus.be.server.reservation.application.dto.ReservationResult;
-import kr.hhplus.be.server.reservation.presentation.dto.response.ReservationResponse;
+import kr.hhplus.be.server.reservation.application.dto.PaymentResult;
 import lombok.Builder;
 
 @Builder
